test: split #933 journey into partials + pin current single-process behavior green
Addresses PR #941 review: the 401-line test file becomes a partial-class
split (green journey / documented-red inventory / .lun round-trip), and the
duplicated StraightPath segment builder moves onto the design class as the
one shared helper.

Folds in the parallel session's findings as green "today" pins beside the
red stations: the process lock really rejects the second chiplet's process
(#935), per-process bend minima resolve when locked while Playground gets
one 10 µm fallback for everything (#937), and the mixed-backend GDS export
really sizes every route with the single majority cross-section xs_nc
(#939) — a future per-chiplet fix turns each pin red as a tripwire.

Duplicate filings #942-#945 closed in favor of #935/#937/#938/#939.
